Rise buffer size for getpwnam_r and getgrnam_r
In some cases, with large /etc/group or /etc/passwd files, calls to
getgrnam_r (in gid_exists) or getpwnam_r (in uid_exists) leads to
errno 34 and fail to start the service.

This fix rises the buffer size but it would be better to rise it
dynamically depending on return value. See man(3) getgrnam_r.